split XCTViewAssert to each of purpose
