+++ Tests/SwiftLayoutTests/Util/SLTAssertConstraintsFunctions.swift
@@ -1,43 +1,14 @@
 //
-//  XCTViewAssert.swift
+//  File.swift
 //  
 //
-//  Created by oozoofrog on 2022/04/04.
+//  Created by oozoofrog on 2022/04/25.
 //
 
 import Foundation
-@testable import SwiftLayout
+import UIKit
 import XCTest
-
-func SLTAssertView(_ view: @autoclosure () throws -> UIView,
-                   superview: @autoclosure () throws -> UIView?,
-                   subviews: @autoclosure () throws -> [UIView]?,
-                   file: StaticString = #filePath, line: UInt = #line) {
-    do {
-        let view = try view()
-        let superview = try superview()
-        let subviews = try subviews() ?? []
-        if view.superview != superview {
-            if let superviewForView = view.superview, let superview = superview {
-                XCTFail("view<\(view.testDescription)>.superview(\(superviewForView.testDescription)) is not equal with superview<\(superview.testDescription)>", file: file, line: line)
-                return
-            } else if let superviewForView = view.superview {
-                XCTFail("view<\(view.testDescription)>.superview<\(superviewForView.testDescription)> is not equal with superview<nil>", file: file, line: line)
-                return
-            } else if let superview = superview {
-                XCTFail("view<\(view.testDescription)>.superview<nil> is not equal with superview<\(superview.testDescription)>", file: file, line: line)
-                return
-            }
-        }
-        let subviewsInView = view.subviews.map(\.testDescription)
-        let subviewsForTest = subviews.map(\.testDescription)
-        if !subviewsInView.elementsEqual(subviewsForTest) {
-            XCTFail("view<\(view.testDescription)>.subviews(\(subviewsInView.count))[\(subviewsInView.joined(separator: ", "))] is not equal with subviews(\(subviewsForTest.count))[\(subviewsForTest.joined(separator: ", "))]", file: file, line: line)
-        }
-    } catch {
-        XCTFail(error.localizedDescription, file: file, line: line)
-    }
-}
+@testable import SwiftLayout
 
 func SLTAssertConstraintsEqual(_ constraints: [NSLayoutConstraint],
                                firstView: UIView, secondView: UIView, tags: [UIView: String] = [:],
@@ -167,29 +138,5 @@ private func typeAndAddressFromView(_ view: UIView) -> String {
 private extension String {
     func replacingRegex(of: String, with: String = "") -> String {
         replacingOccurrences(of: of, with: with, options: .regularExpression, range: startIndex..<endIndex)
-    }
-}
-
-private extension UIAccessibilityIdentification {
-    var testDescription: String {
-        if let identifier = accessibilityIdentifier {
-            return "\(type(of: self)): \(identifier)"
-        } else {
-            return "\(type(of: self)): \(Unmanaged.passUnretained(self).toOpaque())"
-        }
-    }
-}
-
-struct TestAnchors {
-    let constraints: [NSLayoutConstraint]
-    init(first: NSObject, second: NSObject? = nil, @AnchorsBuilder anchors: () -> AnchorsContainer) {
-        self.constraints = anchors().constraints(item: first, toItem: second)
-    }
-}
-
-@resultBuilder
-struct TestAnchorsBuilder {
-    static func buildBlock(_ components: TestAnchors...) -> [NSLayoutConstraint] {
-        components.flatMap(\.constraints)
     }
 }
